Fix Secure Boot with SEV-SNP
The emulated variable store used for Secure Boot uses runtime memory
to store the variables. The memory is allocated  and accessed shortly
after InitializeRamRegions().

This does not work with SEV-SNP, because the runtime memory wasn't
made private nor pvalidated (which happens later on in
AmdSevInitialize()). This tricks the CPU into thinking the access is
some sort of MMIO access, but fails horribly nonetheless and crashes.

Fix it by allocating the variable store after SEV-SNP has been
initialized.

Signed-off-by: Sabin Rapan <sabrapan@amazon.com>